Seperate tests by their page
