+++ app/views/posts/index.html.erb
@@ -11,7 +11,7 @@
       <div class="user-post">
         <%= link_to user_post_path(@user, post) do %>
           <div class="post-card">
-            <h3>Post #<%=index+1%></h3>
+            <h3>Post #<%=index+1%>: <%=post.title%></h3>
             <p>
               <%= post.text.slice(0, 200)... %>
             </p>
